requirements.txt
fastapi
uvicorn[standard]
web3
python-dotenv
httpx
cachetools
orjson
